+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 **Changes**
 - Updated Stripe iOS SDK from 25.0.1 to 25.6.0
-- Updated Stripe Android SDK from 22.2.+ to 22.7.0
+- Updated Stripe Android SDK from 22.2.+ to 22.7.+
 
 ## 0.57.3 - 2026-01-21